mobile ui update
